Advertisements/link to users

+++ Controllers/Advertisements/AdvertisementController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Authorization;
+﻿using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Resallie.Models;
 using Resallie.Services.Advertisements;
@@ -7,7 +8,7 @@ namespace Resallie.Controllers.Advertisements;
 
 [Route("[controller]")]
 [ApiController]
-public class AdvertisementController : ControllerBase
+public class AdvertisementController : BaseController
 {
     private readonly AdvertisementService _service;
 
@@ -22,47 +23,56 @@ public class AdvertisementController : ControllerBase
         return _service.GetAll();
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] Advertisement advertisement)
     {
         if (advertisement.Category != null) return BadRequest();
+
+        advertisement.UserId = GetCurrentUserId();
 
         advertisement = await _service.Create(advertisement);
 
         return Ok(advertisement);
     }
 
-    [Authorize]
     [HttpGet("{id}")]
     public async Task<IActionResult> Get(int id)
     {
         Advertisement? advertisement = await _service.Get(id);
         
         if (advertisement == null) return NotFound();
-
+        
         return Ok(advertisement);
     }
     
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<IActionResult> Delete([FromRoute] int id)
     {
+        bool isOwner = await _service.IsAdvertisementOwner(GetCurrentUserId(), id);
+        if (!isOwner) return Forbid();
+        
         bool success = await _service.Delete(id);
 
-        if (!success) return NotFound();
+        if (!success) return BadRequest();
 
         return Ok();
     }
     
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] Advertisement advertisement)
     {
         if (advertisement.Category != null) return BadRequest();
         
+        bool isOwner = await _service.IsAdvertisementOwner(GetCurrentUserId(), id);
+        if (!isOwner) return Forbid();
+        
         Advertisement? oldAdvertisement = await _service.Get(id);
         if (oldAdvertisement == null) return NotFound();
         
         advertisement = await _service.Update(advertisement, oldAdvertisement);
-
 
         return Ok(advertisement);
     }
